requirements.txt
python-gitlab
pyyaml
Jinja2
